cloud-sync/requirements.txt
paho-mqtt
python-dotenv
azure-eventhub